OPENEUROPA-2252: Moving dependency to base class.
